+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- Fix course and organization code fields normalization and validation
 - Fix search bug due to wrong ordering of char filters
 - Stop indexing unpublished categories
 - Repatriate within our codebase the deprecated bootstrap mixin